Cleanup: Move clickhouse helpers under a separate testing folder
