+++ ci/kubetest/test_clickhouse_persistence_enabled_existing_claim.py
@@ -7,10 +7,10 @@ from utils import (
     NAMESPACE,
     cleanup_k8s,
     exec_subprocess,
-    get_clickhouse_statefulset_spec,
     helm_install,
     wait_for_pods_to_be_ready,
 )
+from helpers.clickhouse import get_clickhouse_statefulset_spec
 
 logging.basicConfig(level=logging.DEBUG)
 log = logging.getLogger()
