+++ ci/kubetest/test_clickhouse_service_loadbalancer.py
@@ -1,6 +1,7 @@
 import pytest
 
-from utils import cleanup_k8s, get_clickhouse_cluster_service_spec, helm_install, wait_for_pods_to_be_ready
+from utils import cleanup_k8s, helm_install, wait_for_pods_to_be_ready
+from helpers.clickhouse import get_clickhouse_cluster_service_spec
 
 HELM_INSTALL_CMD = """
 helm upgrade \
